moved contextMiddleware
