+++ packages/commons/src/index.ts
@@ -2,3 +2,4 @@ export * from "./authentication/index.js";
 export * from "./config/index.js";
 export * from "./errors/index.js";
 export * from "./logging/index.js";
+export * from "./context/index.js";
